+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/lib/puppet/functions/environment/data.rb
