+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/modules/abc/manifests/init.pp
