+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/modules/bad_data/manifests/init.pp
