+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/modules/bca/manifests/init.pp
