+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/modules/empty_json/manifests/init.pp
