+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/modules/empty_json/metadata.json
