+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/modules/empty_key_json/hiera.yaml
