+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/modules/hieraprovider/metadata.json
