+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/modules/meta/manifests/init.pp
