+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/modules/metawcp/manifests/init.pp
