+++ spec/fixtures/unit/functions/lookup_fixture/environments/production/modules/no_provider/manifests/init.pp
